# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Documentation for the Application API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Laravel Server #1"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "summary": "Display a listing of the products.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "products"
                ]
            },
            "post": {
                "summary": "Store a newly created product in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "price",
                                    "brand",
                                    "unit"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "unit": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "products"
                ]
            }
        },
        "/products/{product}": {
            "get": {
                "summary": "Display the specified product.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "products"
                ]
            },
            "put": {
                "summary": "Update the specified resource in product.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "requestBody": {
                    "0": {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "price",
                                    "brand",
                                    "unit"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "unit": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "products"
                ]
            },
            "patch": {
                "summary": "Update the specified resource in product.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "requestBody": {
                    "0": {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code",
                                    "name",
                                    "price",
                                    "brand",
                                    "unit"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "brand": {
                                        "type": "string"
                                    },
                                    "unit": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "products"
                ]
            },
            "delete": {
                "summary": "Remove the specified product from storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "products"
                ]
            }
        },
...
    },
...
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}